backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
aiomysql==0.2.0
aiosqlite==0.20.0
anthropic==0.39.0
sqlglot==25.0.0
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
redis==5.1.1
celery==5.4.0
reportlab==4.2.4
jinja2==3.1.4
websockets==13.0
httpx==0.27.2
python-dotenv==1.0.1
loguru==0.7.2
tenacity==9.0.0
cryptography==43.0.1

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
factory-boy==3.3.1
ruff==0.7.0
mypy==1.12.0
